Improved the nbt load function
Previously if no object was given to the load method it would give a criptic error message. This fixes that.
Previously there were two inputs for the given data. It was a little unobvious which input should be used and that there was a second one (usually I had to try twice to get it right). This squishes both inputs into one input. The second is currently kept for backwards compatabilty but I reccommend we remove it at some point and make the first required.
I have also added some more exceptions.
I have also renamed filename_or_buffer to filepath_or_buffer to better reflect what it is.