Updated to cython 3 and simplified operator methods
The old cython would pass in the first object in a self regardless if it was actually and instance of the class.
Cython 3 adds support for the reverse operator methods making it behave like normal python.
This means we don't need to do primitive conversion on self giving a bit of a speed boost.