Added template files
Many of these methods were previously defined in the base classes where the value attribute was not defined.
Due to the way cython works those attributes were not defined on the cython side and it was falling back to the python definition.
I would like to make the attribute private so that the mutable types cannot be directly modified but that breaks the code in the base class.
The solution is to define all of those methods in each class however that is unmanageable to do manually.
Cython's tempia language allows a template to include other code so that the shared methods are only defined once and can be baked out.