Changed hard coded documentation to runtime documentation
The docstring is now patched at runtime from the source object docstring.
There were issues before that the generated source code would vary based on the version of python it was generated with.
There were also issues that methods could not be conditionally be defined because older versions would try and get their docstring.