Refactored the template logic
The template logic was previously mixed in with the source.
This moves the template logic and files into a separate module to make it easier to manage.
Source templates are in templates/src
inherited templates are in template/templates
Split the tempita module into the wrapper logic and the core tempita logic.
Modified path behaviour to pull from the correct paths.