Simplified the array class a bit
The array class was leaning too much into the numpy array.
If we want we can bring these methods back in the future but I think it would be wise to omit them for now
Split the array logic into each class. Previously this was handled in the base class with subclasses changing the class variables.