Removed equals, strict_equals and is_other
strict_equals is now the same behaviour as the normal equals method.
equals is the same behaviour as the python behaviour so the python object can be used for this.
I thought of a way to cache python objects for easy re-access from the c++ object so this method is no longer needed.