Store TAG_String as bytes internally
As of Bedrock 1.18.30 there exists a key where the TAG_String payload is not a valid utf-8 byte sequence.
The previous implementation of decoding as latin broke because the result encoded to a different byte sequence which was used as a UUID.
The TAG_String payload is now stored as bytes and decoded when accessed by python.
Added py_data which returns the raw bytes. py_data in 2.0 replaces value and I think the raw bytes should be the default way to access the data.
TAG_String now takes str and bytes as input and encodes using utf-8 if the input is a string.