Added non utf-8 byte string support to StringTag
Modified the StringTag class to support byte strings that are not utf-8 compliant.
Bedrock 1.18.30 added string tags containing data that is not utf-8 compliant.
As such the nbt library needs to adapt to be able to read and write this data correctly.
The contains, getitem, iter and len methods have been removed because the return values are ambiguous whether they would deal with bytes or string