Removed closing of the `read`able object
If an object was given to this function that has a read method it should be the job of the calling code to close the object.
There may be cases where the calling code wants to continue using the file handler so forcefully closing it here is not a good idea.