Added customisable string decoder
String decoder can now be specified.
Java uses a modified version of UTF-8 so the encoding scheme needs to be customisable.
Replaced load_name with load_string and made decoding scheme customisable.
I think it is safe to assume that the encoding scheme for names is the same as string tags.