Modified NamedTag getitem
Added support for both the old and new format to the getitem method.
If the item is an int it uses the new behaviour and if it is a string it uses the old behaviour and warns it is depreciated.
An iter method had to be added to make the old iter behaviour work.