Modified load functions
Previously there was one function called load that had different outputs depending on the input which made it a little difficult to work with.
I split this into two functions called load_one and load_many.
I have decided load_one is a pointlessly long name and to keep the load function and make it behave like load_one.
The old behaviour will hang around for a bit to let code catch up.
This means that the change will only effect calls using the count and offset inputs and the rest will not need changing.