Modified typed get and added typed pop
The exception behaviour of typed get methods was confusing.
There is now a boolean flag which switches between returning default on error and raising exceptions on error.
Added typed pop methods that behave the same as the typed gets methods but also removes the key if it exists and the type is correct.
Added tests