Improve build tooling (#108)
* Implement conditional compiler freezing

If "AMULET_FREEZE_COMPILER" config variable is defined as a true value, the runtime requirement of amulet-compiler-version will be frozen and the compiler name and version will be added to the wheel file name as a build tag.
amulet-compiler-version must be pre-built and uploaded to pypi before this can be used.

* Update workflows

* Remove git dependency

pypi does not allow git dependencies

* Build specialised version of compiler version

* Update compiler version

* Create sdist and wheel separately